requirements.txt
aliyun-python-sdk-core-v3==2.13.33
amqp==5.3.1
APScheduler==3.10.1
async-timeout==4.0.2
backports.entry-points-selectable==1.2.0
backports.zoneinfo==0.2.1
bidict==0.22.1
billiard==4.2.1
blinker==1.6.2
celery==5.5.3
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
concurrent_log==1.0.1
cryptography==40.0.2
cymysql==0.9.18
decorator==5.2.1
Deprecated==1.2.13
distlib==0.3.6
et-xmlfile==1.1.0
filelock==3.12.0
Flask==2.2.3
Flask-APScheduler==1.12.4
Flask-Cors==3.0.10
Flask-HTTPAuth==4.8.0
Flask-Login==0.6.2
Flask-Script==2.0.6
Flask-SocketIO==5.3.4
flask-sqlacodegen==1.1.8
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
gevent==22.10.2
gevent-websocket==0.10.1
googleapis-common-protos==1.59.1
greenlet==2.0.2
grpcio==1.70.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.6.0
inflect==6.0.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
kombu==5.5.4
loguru==0.7.0
MarkupSafe==2.1.2
marshmallow==3.22.0
munch==4.0.0
numpy==1.24.3
openpyxl==3.1.2
opentelemetry-api==1.33.1
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-jaeger-proto-grpc==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-flask==0.54b1
opentelemetry-instrumentation-requests==0.54b1
opentelemetry-instrumentation-wsgi==0.54b1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
packaging==23.1
paho-mqtt==1.6.1
pandas==2.0.1
pika==1.3.2
Pillow==9.5.0
pipenv==2023.4.29
platformdirs==3.5.1
portalocker==2.7.0
prompt_toolkit==3.0.51
protobuf==4.25.8
py==1.11.0
pycparser==2.21
pydantic==1.10.7
pyexiv2==2.8.1
PyJWT==2.9.0
PyMySQL==1.0.3
pyparsing==3.0.9
pysnowflake==0.1.3
python-dateutil==2.8.2
python-engineio==4.4.1
python-socketio==5.8.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==311
qiniu==7.10.0
redis==4.5.5
requests==2.30.0
retry==0.9.2
six==1.16.0
SQLAlchemy==2.0.15
thrift==0.22.0
tornado==6.3.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.22
vine==5.1.0
virtualenv==20.23.0
wcwidth==0.2.13
Werkzeug==2.3.4
win32-setctime==1.1.0
wrapt==1.15.0
WTForms==3.0.1
zipp==3.15.0
zope.event==4.6
zope.interface==6.0